Localization system added
Localization system added. Can be tested in Pille-Sandbox.unity. Enter a string in English in the LocalizationTest component on the manager object, set the language to a language code and run. Then press space in the game and read the return value from the console. Make sure that the search word and language code exists in the localizationData.csv found in "Ressources\LocalizationData". NOTE: The system is case-sensitive